+++ src/action/fake_planner/include/fake_planner/fake_planner_node.hpp
@@ -16,9 +16,9 @@
 
 #include <memory>
 #include <string>
-#include <vector>
 
 #include "behaviour_msgs/msg/execute_behaviour.hpp"
+#include "fake_planner/fake_planner_core.hpp"
 #include "nav_msgs/msg/odometry.hpp"
 #include "rclcpp/rclcpp.hpp"
 #include "rclcpp_lifecycle/lifecycle_node.hpp"
@@ -48,19 +48,7 @@ public:
   CallbackReturn on_shutdown(const rclcpp_lifecycle::State & state) override;
 
 private:
-  // Parses a maneuver JSON (segment list), expands the segments into a smooth path and resamples
-  // it to uniform spacing in wp_x_/wp_y_/wp_speed_. Returns false and sets `error` on any I/O,
-  // JSON, or schema failure.
-  bool loadManeuver(const std::string & path, std::string & error);
-  // Builds full_traj_ from the loaded waypoints, applying the current anchor transform.
-  void buildTrajectory();
-  // Slices the rolling window out of full_traj_ around the vehicle's current position into
-  // trajectory_, mirroring what the real planner emits (a short horizon, not the whole route).
-  void updateWindow();
-  // Index of the full-path point closest to (veh_x_, veh_y_), searched forward from the last
-  // match so a self-intersecting maneuver doesn't snap back to an earlier lap.
-  std::size_t nearestIndex() const;
-  // Publishes the trajectory as a MarkerArray, mirroring trajectory_planner's visualization
+  // Publishes the current window as a MarkerArray, mirroring trajectory_planner's visualization
   // (speed-sized spheres + speed labels on `trajectory_markers`, only when subscribed).
   void publishMarkers();
   void timerCallback();
@@ -74,6 +62,9 @@ private:
   void resetTrajectory(
     const std_srvs::srv::Trigger::Request::SharedPtr request, std_srvs::srv::Trigger::Response::SharedPtr response);
 
+  // Maneuver expansion and rolling-window slicing — holds no ROS state
+  std::unique_ptr<FakePlannerCore> core_;
+
   // Parameters
   std::string trajectory_topic_;
   std::string behaviour_topic_;
@@ -86,35 +77,16 @@ private:
   bool start_on_activate_{true};
   std::string maneuver_file_;
   std::string marker_topic_;
-  double horizon_m_{30.0};
+  double horizon_m_{35.0};
   double trail_m_{2.0};
   double respawn_jump_m_{5.0};
-
-  // Set from the maneuver's "closed" flag: a closed circuit laps forever (the window wraps
-  // past the end), an open maneuver ends in a stop.
-  bool closed_{false};
-
-  // Waypoints expanded from the maneuver segments (parallel arrays; relative to the anchor pose).
-  std::vector<double> wp_x_;
-  std::vector<double> wp_y_;
-  std::vector<double> wp_speed_;
-
-  // The whole anchored maneuver (built once the anchor pose is known), and the rolling window
-  // of it that actually gets published each tick.
-  wato_trajectory_msgs::msg::Trajectory full_traj_;
-  wato_trajectory_msgs::msg::Trajectory trajectory_;
-  bool trajectory_ready_{false};
-  std::size_t window_start_{0};
 
   // Gates publishing: when false the node stays silent so the controller holds in standby.
   // Toggled by the start/stop services; seeded from start_on_activate on activation.
   bool trajectory_started_{true};
 
-  // Anchor pose (SE(2), in frame_id_) that the waypoints are laid out from.
+  // Whether the maneuver has been laid out from a pose yet (see anchor_to_first_pose_).
   bool anchored_{false};
-  double anchor_x_{0.0};
-  double anchor_y_{0.0};
-  double anchor_yaw_{0.0};
 
   // Latest vehicle position in frame_id_, tracked continuously to slide the window.
   bool have_pose_{false};
