--- src/action/fake_planner/launch/controllers/pure_pursuit.launch.yaml
@@ -1,76 +1,0 @@
-# controllers/pure_pursuit.launch.yaml -- ackermann pure pursuit as the controller under test.
-#
-# Included by fake_planner.launch.yaml via controller:=pure_pursuit (the default); one file per
-# controller so `controller:=` can pick the file by name. Also brings the lifecycle manager,
-# which lives here because it must list the controller node by its concrete name.
-
-launch:
-  - arg:
-      name: odom_topic
-      default: "/ego/odom"
-      description: Odometry source for the controller
-
-  - arg:
-      name: ackermann_topic
-      default: "/carla/ackermann_control/command"
-      description: Where the controller publishes AckermannDriveStamped commands
-
-  - arg:
-      name: standby_speed
-      default: "-0.5"
-      description: Controller speed in standby. Package default -0.5 reverses; use 0.0 to hold still.
-
-  - arg:
-      name: use_sim_time
-      default: "false"
-      description: Use /clock. Required in CARLA, else the controller can't transform the trajectory.
-
-  - arg:
-      name: autostart
-      default: "true"
-      description: Automatically configure and activate managed nodes
-
-  # Pure Pursuit Controller (under test)
-  - node:
-      pkg: "ackermann_pure_pursuit"
-      exec: "pure_pursuit_node"
-      name: "pure_pursuit_node"
-      namespace: "action"
-      output: "screen"
-      param:
-        - from: "$(find-pkg-share ackermann_pure_pursuit)/config/params.yaml"
-        - name: "standby_speed"
-          value: "$(var standby_speed)"
-        - name: "use_sim_time"
-          value: "$(var use_sim_time)"
-      remap:
-        # Subscribers
-        - from: "trajectory"
-          to: "/action/fake_planner/trajectory"
-        - from: "execute_behaviour"
-          to: "/action/fake_planner/execute_behaviour"
-        - from: "odom"
-          to: "$(var odom_topic)"
-        # Publishers
-        - from: "ackermann"
-          to: "$(var ackermann_topic)"
-
-  # Lifecycle Manager
-  - node:
-      pkg: "wato_lifecycle_manager"
-      exec: "wato_lifecycle_manager_node"
-      name: "fake_planner_lifecycle_manager"
-      namespace: "action"
-      param:
-        - name: "node_names"
-          value:
-            - /action/fake_planner/fake_planner_node
-            - /action/pure_pursuit_node
-        - name: autostart
-          value: $(var autostart)
-        - name: transition_timeout_s
-          value: 10.0
-        - name: bond_timeout_s
-          value: 4.0
-        - name: bond_enabled
-          value: false
