--- src/action/fake_planner/launch/oval_track.launch.yaml
@@ -1,85 +1,0 @@
-# oval_track.launch.yaml -- fake planner on the oval track: replays maneuvers/oval_track.json
-# so the controller under test drives continuous laps. This is fake_planner.launch.yaml with
-# the oval maneuver and anchor_to_first_pose:=false (the maneuver is in absolute track coords,
-# so it must not be anchored to the spawn pose). The oval MAP (world_model) lives in a separate
-# container and is optional -- the fake planner doesn't consume it.
-#
-# Requires CARLA running the oval world (supplies /ego/odom + TF, takes the ackermann cmd):
-#     watod up   # with CARLA_SCENARIO=oval in watod-config.local.sh
-#     ros2 launch fake_planner oval_track.launch.yaml
-#   Pick the controller:   controller:=pure_pursuit (default) or controller:=mpc
-#   Watch the trajectory:  /action/fake_planner/trajectory_markers (MarkerArray, same format
-#                          as the real trajectory_planner), published even while holding.
-#   Real vehicle: odom_topic:=/world_modeling/transform/odometry ackermann_topic:=/action/ackermann
----
-launch:
-  - arg:
-      name: odom_topic
-      default: "/ego/odom"
-      description: Odometry source (controller speed; anchoring is off for this maneuver)
-
-  - arg:
-      name: ackermann_topic
-      default: "/carla/ackermann_control/command"
-      description: Where the controller publishes AckermannDriveStamped commands
-
-  - arg:
-      name: frame_id
-      default: "odom"
-      description: Frame the trajectory is published in (must be TF-connected to base_footprint)
-
-  - arg:
-      name: controller
-      default: "pure_pursuit"
-      description: "Controller under test: pure_pursuit or mpc"
-
-  - arg:
-      name: maneuver
-      default: "oval_track"
-      description: Maneuver to replay. Anchoring is off here, so it must carry an absolute "start" pose on the track.
-
-  - arg:
-      name: autostart
-      default: "true"
-      description: Automatically configure and activate managed nodes
-
-  - arg:
-      name: start_on_activate
-      default: "false"
-      description: Held in standby by default; call the start_trajectory service to drive.
-
-  - arg:
-      name: use_sim_time
-      default: "true"
-      description: CARLA stamps odom/TF with sim time; set false on the real vehicle.
-
-  - arg:
-      name: standby_speed
-      default: "0.0"
-      description: Hold still (0.0) in standby; the controller's -0.5 default reverses off the track.
-
-  # Fake planner (oval maneuver, absolute coords) + controller + lifecycle manager.
-  - include:
-      file: $(find-pkg-share fake_planner)/launch/fake_planner.launch.yaml
-      arg:
-        - name: maneuver
-          value: $(var maneuver)
-        - name: controller
-          value: $(var controller)
-        - name: odom_topic
-          value: $(var odom_topic)
-        - name: ackermann_topic
-          value: $(var ackermann_topic)
-        - name: frame_id
-          value: $(var frame_id)
-        - name: autostart
-          value: $(var autostart)
-        # The oval maneuver carries an absolute "start" pose on the track lane.
-        - name: anchor_to_first_pose
-          value: "false"
-        - name: start_on_activate
-          value: $(var start_on_activate)
-        - name: standby_speed
-          value: $(var standby_speed)
-        - name: use_sim_time
-          value: $(var use_sim_time)
