+++ src/simulation/carla_ros_bridge/carla_scenarios/carla_scenarios/scenarios/oval_track_scenario.py
@@ -34,7 +34,8 @@ class OvalTrackScenario(ScenarioBase):
     replays as a maneuver.
 
     Pair with:
-        ros2 launch fake_planner oval_track.launch.yaml
+        ros2 launch fake_planner fake_planner.launch.yaml maneuver:=oval_track
+            anchor_to_first_pose:=false start_on_activate:=false
     """
 
     # Generated by tools/generate_oval_track.py; mounted into the container from maps/.
