--- src/world_modeling/world_modeling_bringup/launch/world_modeling_oval_map_only.launch.yaml
@@ -1,97 +1,0 @@
-# Copyright (c) 2025-present WATonomous. All rights reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-# world_modeling_oval_map_only.launch.yaml -- loads + visualizes the oval lanelet map
-# (maps/osm/oval_track_4_lane.osm) via world_model + map markers. Three nodes, no behaviour,
-# costmap, prediction or SLAM.
-#
-# This is the bringup to pair with the fake planner: the fake planner publishes its own
-# trajectory and its own execute_behaviour heartbeat, so it consumes nothing from world_modeling
-# and needs only /ego/odom + TF from CARLA. The map here is purely visual -- it puts the lane
-# geometry under the trajectory markers in Foxglove. Nothing includes this file; run it yourself
-# alongside fake_planner/launch/oval_track.launch.yaml, or skip it entirely.
-#
-# For the real planning chain on the oval (behaviour -> lattice -> trajectory_planner), use
-# world_modeling_oval.launch.yaml instead -- that is the full 15-node stack.
-#
-#   ros2 launch world_modeling_bringup world_modeling_oval_map_only.launch.yaml
----
-launch:
-  - arg:
-      name: config_file
-      default: $(find-pkg-share world_modeling_bringup)/config/world_model_oval.yaml
-      description: world_model configuration file (oval map path + projector)
-
-  - arg:
-      name: autostart
-      default: "true"
-      description: Automatically configure and activate world_model
-
-  # World Model -- loads the oval OSM
-  - node:
-      pkg: world_model
-      exec: world_model_node
-      name: world_model
-      namespace: "world_modeling"
-      output: screen
-      param:
-        - from: $(var config_file)
-      remap:
-        - from: "world_objects"
-          to: "world_objects_enriched"
-        - from: "lane_context"
-          to: "lanelet/lane_context"
-        - from: "map_visualization"
-          to: "lanelet/map_visualization"
-        - from: "route_ahead"
-          to: "lanelet/route_ahead"
-        - from: "lanelet_ahead"
-          to: "lanelet/lanelet_ahead"
-
-  # Lanelet map visualization markers
-  - node:
-      pkg: lanelet_markers
-      exec: map_viz_markers_node
-      name: map_viz_markers_node
-      namespace: "world_modeling"
-      output: screen
-      param:
-        - from: $(var config_file)
-      remap:
-        - from: "lane_context"
-          to: "lanelet/lane_context"
-        - from: "map_visualization"
-          to: "lanelet/map_visualization"
-        - from: "markers"
-          to: "lanelet/map_viz_markers"
-
-  # Lifecycle Manager
-  - node:
-      pkg: wato_lifecycle_manager
-      exec: wato_lifecycle_manager_node
-      name: world_model_oval_lifecycle_manager
-      namespace: "world_modeling"
-      output: screen
-      param:
-        - name: node_names
-          value:
-            - /world_modeling/world_model
-        - name: autostart
-          value: $(var autostart)
-        - name: transition_timeout_s
-          value: 10.0
-        - name: bond_timeout_s
-          value: 4.0
-        - name: bond_enabled
-          value: false
