+++ watod_scripts/watod-setup-env.sh
@@ -159,8 +159,6 @@ CLAUDE_CODE=${CLAUDE_CODE:-"true"}
 
 ################################  Simulation  ########################################
 CARLA_RENDER_MODE=${CARLA_RENDER_MODE:-"no_gpu"}
-# Scenario the scenario_server loads on startup (override via CARLA_SCENARIO).
-CARLA_SCENARIO=${CARLA_SCENARIO:-"carla_scenarios.scenarios.light_traffic_scenario"}
 # Enable pygame HUD when running without GPU (provides web-based visualization fallback)
 if [[ "$CARLA_RENDER_MODE" == "no_gpu" ]]; then
   PYGAME_HUD_ENABLED=${PYGAME_HUD_ENABLED:-"true"}
@@ -210,7 +208,6 @@ append "CLAUDE_CODE" "$CLAUDE_CODE"
 # Simulation
 append "CARLA_RENDER_MODE" "$CARLA_RENDER_MODE"
 append "PYGAME_HUD_ENABLED" "$PYGAME_HUD_ENABLED"
-append "CARLA_SCENARIO" "$CARLA_SCENARIO"
 
 append "REGISTRY" "$REGISTRY"
 append "REPOSITORY" "$REPOSITORY"
